+++ lib/widgets/cart_item.dart
